% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtopread.R
\name{readAreas}
\alias{readAreas}
\title{help file for creating \code{sf} objects with observations and/or
predictionLocations of data with a spatial support}
\usage{
readAreas(object, adir = ".", ftype = "xy", projection = NA, ...)
}
\arguments{
\item{object}{either name of file with areal information or data frame with
observations}

\item{adir}{directory where the files with areal information are to be found}

\item{ftype}{type of file, the only type supported currently is "xy",
referring to x- and y-coordinates of boundaries}

\item{projection}{add projection to the object if input is boundary-files}

\item{...}{further parameters to be passed to \code{\link{readAreaInfo}}}
}
\value{
The function creates an \code{sf} object of observations and/or
predictionLocations, depending on the information given in \code{object}.
and/or predictionLocations, depending on the information given in
\code{object}.
}
\description{
readAreas will read area-files, add observations and convert the result to
\code{\link[sf]{sf}}.
}
\details{
If \code{object} is a file name, \code{\link{readAreaInfo}} will be called.
If it is a data frame with observations and/or predictionLocations, the
function will read areal data from files according to the ID associated with
from files according to the ID associated with each
observation/predictionLocation.

The function is of particular use when data are not available as
shape-files, or when the observations are not part of the shape-files. This
function is mainly for compatibility with the former FORTRAN-version. The
simplest way to read the data in that case is through
\code{\link[sf]{st_read}}. See also \code{\link{utop-package}}.
}
\author{
Jon Olav Skoien
}
\keyword{spatial}
