% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtopKrige.R, R/tskrige.R, R/zzz.R
\name{rtopKrige.rtop}
\alias{rtopKrige.rtop}
\alias{rtopKrige.default}
\alias{rtopKrige.stars}
\alias{rtopKrige}
\title{Spatial interpolation of data with spatial support}
\usage{
\method{rtopKrige}{rtop}(object, varMatUpdate = FALSE, params = list(), ...)

\method{rtopKrige}{default}(
  object,
  predictionLocations = NULL,
  varMatObs,
  varMatPredObs,
  varMat,
  params = list(),
  formulaString,
  sel,
  wret = FALSE,
  dObs = NULL,
  dPred = NULL,
  trendObs = NULL,
  trendPred = NULL,
  ...
)

\method{rtopKrige}{stars}(
  object,
  predictionLocations = NULL,
  varMatObs,
  varMatPredObs,
  varMat,
  params = list(),
  formulaString,
  sel,
  olags = NULL,
  plags = NULL,
  lagExact = TRUE,
  ...
)

rtopKrige(object, ...)
}
\arguments{
\item{object}{object of class \code{rtop}, \code{\link[sf]{sf}}, or
\code{stars}}

\item{varMatUpdate}{logical; if TRUE, also existing variance matrices will
be recomputed, if FALSE, only missing variance matrices will be computed,
see also \code{\link{varMat}}}

\item{params}{a set of parameters, used to modify the default parameters for
the \code{rtop} package, set in \code{\link{getRtopParams}}. Additionally,
it is possible overrule some of the parameters in \code{object$params} by
passing them as separate arguments.}

\item{...}{from \code{rtopKrige.rtop}, arguments to be passed to
\code{rtopKrige.default}. In \code{rtopKrige.default}, parameters for
modification of the object parameters or default parameters.  Of particular
interest are \code{cv}, a logical for doing cross-validation, \code{nmax},
and \code{maxdist} for maximum number of neighbours and maximum distance to
neighbours, respectively, and \code{wlim}, the limit for the absolute values
of the weights. It can also be useful to set \code{singularSolve} if some of
the areas are almost similar, see also details below.}

\item{predictionLocations}{\code{\link[sf]{sf}} or
\code{stars} object with prediction locations. NULL if cross
validation is to be performed.}

\item{varMatObs}{covariance matrix of observations, where diagonal must
consist of internal variance, typically generated from call to
\code{\link{varMat}}}

\item{varMatPredObs}{covariance matrix between observation locations and
prediction locations, typically generated from call to \code{\link{varMat}}}

\item{varMat}{list covariance matrices including the two above}

\item{formulaString}{formula that defines the dependent variable as a linear
model of independent variables, see e.g. \code{\link{createRtopObject}} for
more details.}

\item{sel}{array of prediction location numbers, if only a limited number of
locations are to be interpolated/crossvalidated}

\item{wret}{logical; if TRUE, return a matrix of weights instead of the
predictions, useful for batch processing of time series, see also details}

\item{dObs}{list of discretisation points for the observation areas from
\code{\link{rtopDisc}}, used for block-averaging of coordinate-based trend
basis functions when \code{ukTrendSupport = "block"}; computed internally
when needed and not supplied}

\item{dPred}{list of discretisation points for the prediction areas, see
\code{dObs}}

\item{trendObs}{matrix of trend basis functions evaluated at the
observation supports (one row per observation, including the intercept);
mainly for internal use, computed from \code{formulaString} when missing}

\item{trendPred}{matrix of trend basis functions evaluated at the
prediction supports, see \code{trendObs}}

\item{olags}{A vector describing the relative lag which should be applied
for the observation locations. See also details}

\item{plags}{A vector describing the relative lag which should be applied
for the predicitonLocations. See also details}

\item{lagExact}{logical; whether differences in lagtime should be computed
exactly or approximate}
}
\value{
If called with \code{\link[sf]{sf}}, the function returns an
\code{sf} object with predictions, either at the locations defined in
\code{predictionLocations}, or as leave-one-out cross-validation predictions
at the same locations as in object if \code{cv = TRUE}. For
\code{stars}, a stars object with prediction attributes is
returned.

If called with an rtop-object, the function returns the same object with the
predictions added to the object.
}
\description{
rtopKrige perform spatial interpolation or cross validation of data with
areal support.
}
\details{
This function is the interpolation routine of the utop-package. The
simplest way of calling the function is with an rtop-object that contains
the fitted variogram model and all the other necessary data (see
\code{\link{createRtopObject}} or \code{\link{utop-package}}).

The function will, if called with covariance matrices between observations
and between observations and prediction locations, use these for the
interpolation. If the function is called without these matrices,
\code{\link{varMat}} will be called to create them. These matrices can
therefore be reused if necessary, an advantage as it is computationally
expensive to create them.

The interpolation that takes part within \code{rtopKrige.default} is based
on the semivariance matrices between observations and between observations
and prediction locations. It is therefore possible to use this function also
to interpolate data where the matrices have been created in other ways, e.g.
based on distances in physiographical space or distances along a stream.

The function returns the weights rather than the predictions if \code{wret =
TRUE}. This is useful for batch processing of time series, e.g. once the
weights are created, they can be used to compute the interpolated values for
each time step.

rtop is able to take some advantage of multiple CPUs, which can be invoked
with the parameter \code{nclus}. When it gets a number larger than one,
\code{rtopKrige} will start a cluster with \code{nclus} workers, if the
\code{\link{parallel}}-package has been installed.

The parameter \code{singularSolve} can be used when some areas are almost
completely overlapping. In this case, the discretization of them might be
equal, and the covariances to other areas will also be equal. The kriging
matrix will in this case be singular. When \code{singularSolve = TRUE},
\code{rtopKrige} will remove one of the neighbours, and instead work with
the mean of the two observations. An overview of removed neighbours can be
seen in the resulting object, under the name \code{removed}.

Kriging of time series is possible when \code{observations} and
\code{predictionLocations} are \code{stars} vector data cubes
with one spatial \code{sfc} dimension and one time dimension. The
interpolation is still spatial, in the sense that regularisation of the
variograms uses the spatial extent of the observations, not a possible
temporal extent, such as done by Skoien and Bloschl (2007). However, it is
possible to make predictions based on observations from different time steps,
through the use of the lag-vectors. These vectors describe a typical
"delay" for each observation and prediction location. This delay could for
runoff related variables be similar to travel time to each gauging location.
For a certain prediction location, earlier time steps would be picked for
neighbours with shorter travel time and later time steps for neighbours with
slower travel times.

The lagExact parameter indicates whether to use a weighted average of two
time steps, or just the time step which is closest to the difference in lag
times.

Universal kriging is invoked by giving \code{formulaString} a non-trivial
right hand side, e.g. \code{obs ~ elev} or \code{obs ~ x + y}, where the
trend variables are attribute columns of the observations and prediction
locations and/or the reserved coordinate names \code{x} and \code{y}. The
kriging system is then augmented with the trend basis functions instead of
only the unbiasedness constraint of ordinary kriging. The basis functions
are evaluated at the centroid of each support area, or block-averaged over
the discretisation points of \code{\link{rtopDisc}}, controlled by the
parameter \code{ukTrendSupport} (see \code{\link{getRtopParams}}). The
sample variogram is in this case computed from the residuals against an
OLS fit of the trend. For spatiotemporal interpolation the trend variables
have to be time-invariant stars attributes; time-varying stars attributes
are not used as support-level trend covariates.

The use of lag times should in theory increase the computation time, but
might, due to different computation methods, even speed up the computation
when the number of neighbours to be used (parameter nmax) is small compared
to the number of observations. If computation is slow, it can be useful to
test olags = rep(0, \code{dim(observations)[1]}) and similar for
predictionLocations.
}
\examples{

\donttest{
# The following command will download  the complete example data set
# downloadRtopExampleData()
# observations$obs = observations$QSUMMER_OB/observations$AREASQKM

rpath <- system.file("extdata",package="utop")
library(sf)
observations <- sf::st_read(rpath, "observations")
predictionLocations <- sf::st_read(rpath,"predictionLocations")

# Setting some parameters; nclus > 1 will start a cluster with nclus
# workers for parallel processing
params <- list(gDist = TRUE, cloud = FALSE, nclus = 1, rresol = 25)

# Create a column with the specific runoff:
observations$obs <- observations$QSUMMER_OB/observations$AREASQKM

# Build an object
rtopObj <- createRtopObject(observations, predictionLocations,
                           params = params)

# Fit a variogram (function also creates it)
rtopObj <- rtopFitVariogram(rtopObj)

# Predicting at prediction locations
rtopObj <- rtopKrige(rtopObj)

# Cross-validation
rtopObj <- rtopKrige(rtopObj,cv=TRUE)
cor(rtopObj$predictions$observed,rtopObj$predictions$var1.pred)
}

}
\references{
Skoien J. O., R. Merz, and G. Bloschl. Top-kriging -
geostatistics on stream networks. Hydrology and Earth System Sciences,
10:277-287, 2006.

Skoien, J. O. and G. Bloschl. Spatio-Temporal Top-Kriging of Runoff Time
Series. Water Resources Research 43:W09419, 2007.

Skoien, J. O., Bloschl, G., Laaha, G., Pebesma, E., Parajka, J., Viglione,
A., 2014. Rtop: An R package for interpolation of data with a variable
spatial support, with an example from river networks. Computers &
Geosciences, 67.
}
\author{
Jon Olav Skoien
}
\keyword{spatial}
