% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtopVario.R, R/zzz.R
\name{rtopVariogram.sf}
\alias{rtopVariogram.sf}
\alias{rtopVariogram.rtop}
\alias{rtopVariogram.stars}
\alias{rtopVariogram}
\title{create variogram for data with spatial support}
\usage{
\method{rtopVariogram}{sf}(
  object,
  formulaString,
  params = list(),
  cloud,
  discPoints = NULL,
  ...
)

\method{rtopVariogram}{rtop}(object, params = list(), ...)

\method{rtopVariogram}{stars}(
  object,
  formulaString,
  params = list(),
  cloud,
  abins,
  dbins,
  discPoints = NULL,
  ...
)

rtopVariogram(object, ...)
}
\arguments{
\item{object}{object of class \code{rtop} (see \link{utop-package}), an
\code{\link[sf]{sf}} object with areal observations, or a vector data cube
with \code{stars}. Observation data must include an
\code{area} column/attribute or have polygon geometries from which area can
be computed.}

\item{formulaString}{formula that defines the dependent variable as a linear
model of independent variables; suppose the dependent variable has name
\code{z}, for ordinary and simple kriging use the formula \code{z~1}; for
universal kriging, suppose \code{z} is linearly dependent on \code{x} and
\code{y}, use the formula \code{z~x+y}. The formulaString defaults to
\code{"value~1"} if \code{value} is a part of the data set.  If not, the
first column of the data set is used.}

\item{params}{a set of parameters, used to modify the default parameters for
the \code{rtop} package, set in \code{\link{getRtopParams}}.}

\item{cloud}{logical; if TRUE, calculate the semivariogram cloud, can be
used to overrule the cloud parameter in params.}

\item{discPoints}{list of discretisation points of the observation areas
from \code{\link{rtopDisc}}, only used for block-averaging of
coordinate-based trend basis functions when the formula describes
universal kriging and \code{ukTrendSupport = "block"}; computed internally
when needed and not supplied}

\item{...}{parameters to other functions called, e.g. gstat's
\code{\link[gstat]{variogram}}-function}

\item{abins}{possibility to set areal bins (not yet implemented)}

\item{dbins}{possibility to set distance bins (not yet implemented)}
}
\value{
The function creates a variogram, either of type
\code{rtopVariogram} or \code{rtopVariogramCloud}. This variogram is based
on the \code{\link[gstat]{variogram}} function from gstat, but has
additional information about the spatial size or length of the observations.
An rtop-object with the variogram added is returned if the function is
called with an rtop-object as argument.

For spatiotemporal \code{stars} vector data cubes, the variogram is spatial
and averaged over all time steps.
}
\description{
rtopVariogram will create binned variogram or cloud variogram of data with
an areal support.
}
\note{
The variogram cloud is similar to the variogram cloud from
\code{\link[gstat]{gstat}}, with the area/length added to the resulting
data.frame. The binned variogram is also based on the area or length, in
addition to the distance between observations. The bins equally distanced in
the log10-space of the distances and areas (lengths). The size of the bins
is decided from the parameters \code{amul} and \code{dmul}, defining the
number of bins per order of magnitude (1:10, 10:100, and so on).

The distances between areas are in this function based on the centre of
gravity.
}
\examples{

\dontrun{
library(sf)
rpath <- system.file("extdata",package="utop")
observations <- sf::st_read(rpath,"observations")
# Create a column with the specific runoff:
observations$obs <- observations$QSUMMER_OB/observations$AREASQKM

vario <- rtopVariogram(observations, cloud = TRUE)
}

}
\references{
Skoien J. O., R. Merz, and G. Bloschl. Top-kriging -
geostatistics on stream networks. Hydrology and Earth System Sciences,
10:277-287, 2006.

Skoien, J. O., Bloschl, G., Laaha, G., Pebesma, E., Parajka, J., Viglione,
A., 2014. Rtop: An R package for interpolation of data with a variable
spatial support, with an example from river networks. Computers &
Geosciences, 67.
}
\author{
Jon Olav Skoien
}
\keyword{spatial}
