% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utop-package.R
\docType{package}
\name{utop-package}
\alias{utop}
\alias{utop-package}
\title{A package providing methods for analysis and spatial interpolation of data
with an irregular support}
\description{
This package provides geostatistical methods for analysis and interpolation
of data that has an irregular support, such as as runoff characteristics or
population health data. The methods in this package are based on the
top-kriging approach suggested in Skoien et al (2006), with some
extensions from Gottschalk (1993).
}
\section{Workflow}{


The work flow within the package suggests that the user is interested in a
prediction of a process at a series of locations where observations have not
been made. The example below shows a regionalization of mean annual runoff
in Austria.

Although it is possible to perform each step with all necessary arguments,
the easiest interface to the method is to store all variables (such as
observations, prediction locations and parameters) in an rtop-object, which
is created by a call to \code{\link{createRtopObject}}. The element
\code{params} below consists of changes to the default parameters. A further
description can be found in \code{\link{getRtopParams}}. The changes below
means that the functions will use geostatistical distance instead of full
regularization, and that the variogram model will be fitted to the variogram
cloud. Most other functions in the \code{utop}-package can take this
object as an argument, and will add the results as one or more new
element(s) to this object.

The data in the example below are stored as shape-files in the
extdata-directory of the utop-package, use the directory of your own data
instead. The observations consist of mean summer runoff from 138 catchments
in Upper Austria. The predictionLocations are 863 catchments in the same
region. observations and predictionLocations should be stored as
\code{\link[sf]{sf}} polygons. Spatiotemporal data are represented as
vector data cubes with \code{stars}.

\preformatted{ rpath = system.file("extdata",package="utop") library(sf)
observations = st_read(rpath, "observations") predictionLocations =
st_read(rpath,"predictionLocations")

# Create a column with the specific runoff: observations$obs =
observations$QSUMMER_OB/observations$AREASQKM params = list(gDist = TRUE,
cloud = TRUE) rtopObj = createRtopObject(observations, predictionLocations,
params = params) }

There are help-methods available in cases when data are not available as
shape-files, or when the observations are not part of the shape-files. See
\code{\link{readAreaInfo}} and \code{\link{readAreas}}.

A call to \code{\link{rtopVariogram}} adds the sample variogram to the
object, whereas \cr \code{\link{rtopFitVariogram}} fits a variogram model.
The last function will call \code{\link{rtopVariogram}} if \code{rtopObj}
does not contain a sample variogram.

\preformatted{ rtopObj = rtopVariogram(rtopObj) rtopObj =
rtopFitVariogram(rtopObj, maxn = 2000) }

The function \code{\link{checkVario}} is useful to produce some diagnostic
plots for the sample variogram and the fitted variogram model.

\preformatted{ checkVario(rtopObj) }

The interpolation function (\code{\link{rtopKrige}}) solves the kriging
system based on the computed regularized semivariances. The covariance
matrices are created in a separate regularization function
(\code{\link{varMat}}), and are stored in the rtop-object for easier access
if it is necessary to redo parts of the analysis, as this is the
computationally expensive part of the interpolation. Cross-validation can be
called with the argument \code{cv=TRUE}, either in \code{params} or in the
call to \code{\link{rtopKrige}}.

\preformatted{ rtopObj = rtopKrige(rtopObj)
# plot predictions with sf/ggplot2
ggplot(rtopObj$predictions) + aes(fill = var1.pred) + geom_sf()

rtopObj = rtopKrige(rtopObj, cv = TRUE)
ggplot(rtopObj$predictions) + aes(fill = var1.var) + geom_sf() }
}

\references{
L. Gottschalk. Interpolation of runoff applying objective
methods. Stochastic Hydrology and Hydraulics, 7:269-281, 1993.

Skoien J. O., R. Merz, and G. Bloschl. Top-kriging - geostatistics on stream
networks. Hydrology and Earth System Sciences, 10:277-287, 2006.

Skoien, J. O., Bloschl, G., Laaha, G., Pebesma, E., Parajka, J., Viglione,
A., 2014. Rtop: An R package for interpolation of data with a variable
spatial support, with an example from river networks. Computers &
Geosciences, 67.
}
\author{
\strong{Maintainer}: Jon Olav Skøien \email{jon.skoien@gmail.com}

Authors:
\itemize{
  \item Jon Olav Skøien \email{jon.skoien@gmail.com}
  \item Anatoly Tsyplenkov \email{sharp.couch9400@fastmail.com}
}

}
\keyword{spatial}
