% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-disc.R
\name{utop_disc}
\alias{utop_disc}
\title{Discretise spatial supports}
\usage{
utop_disc(object, ...)
}
\arguments{
\item{object}{A \link{Utop} object, spatial object, or variogram object.}

\item{...}{Method-specific arguments. For \link{Utop} objects, \code{params} updates
parameters.}
}
\value{
Discretisation points or an updated \link{Utop} object.
}
\description{
Discretise spatial supports
}
