% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-api.R
\name{utop_krige}
\alias{utop_krige}
\title{Krige with a utop object}
\usage{
utop_krige(object, ...)
}
\arguments{
\item{object}{A fitted \link{Utop} object.}

\item{...}{Arguments passed to the kriging implementation.}
}
\value{
A \link{Utop} object with predictions attached.
}
\description{
Krige with a utop object
}
