% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-params.R
\name{utop_params}
\alias{utop_params}
\title{Create utop parameters}
\usage{
utop_params(..., observations = NULL, formula = NULL)
}
\arguments{
\item{...}{Canonical utop parameter values supplied as named arguments.}

\item{observations}{Optional observations used to infer starting values.}

\item{formula}{Optional model formula.}
}
\value{
A \link{UtopParams} object.
}
\description{
Create utop parameters
}
