% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-sim.R
\name{utop_sim}
\alias{utop_sim}
\title{Simulate with a utop object}
\usage{
utop_sim(object, ...)
}
\arguments{
\item{object}{A \link{Utop} object.}

\item{...}{Method-specific arguments. For \link{Utop} objects, \code{var_mat_update}
and \code{params} control matrix recomputation and parameter updates.}
}
\value{
A \link{Utop} object with simulations attached.
}
\description{
Simulate with a utop object
}
