% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-variogram-model.R
\name{utop_update_variogram}
\alias{utop_update_variogram}
\title{Update a variogram model}
\usage{
utop_update_variogram(object, ...)
}
\arguments{
\item{object}{A \link{Utop} or \link{UtopVariogramModel} object.}

\item{...}{Variogram parameter updates.}
}
\value{
An updated object.
}
\description{
Update a variogram model
}
