% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-s7-classes.R
\name{utop-s7-classes}
\alias{utop-s7-classes}
\alias{Utop}
\alias{UtopParams}
\alias{UtopVariogram}
\alias{UtopVariogramCloud}
\alias{UtopVariogramModel}
\title{S7 classes for utop objects}
\usage{
UtopParams(
  model = "Ex1",
  nugget = FALSE,
  unc = TRUE,
  r_resol = 100,
  h_resol = 5,
  rs_type = "rtop",
  hs_type = "regular",
  cloud = FALSE,
  amul = 2,
  dmul = 3,
  fit_method = 9,
  g_dist_est = FALSE,
  g_dist_pred = FALSE,
  var_clean = FALSE,
  max_dist = Inf,
  n_max = 10,
  n_clus = 1,
  cn_areas = 100,
  clus_type = NULL,
  outfile = NULL,
  partial_overlap = FALSE,
  wlim = 1.5,
  wlim_method = "all",
  singular_solve = FALSE,
  uk_trend_support = "centroid",
  cv = FALSE,
  debug_level = 0,
  par_init = NULL
)

UtopVariogramModel(
  model = "Ex1",
  params = c(1, 1, 0, 0, 1),
  ss_err = NULL,
  criterion = NULL
)

UtopVariogram(data = data.frame())

UtopVariogramCloud(data = data.frame())

Utop(
  observations = NULL,
  prediction_locations = NULL,
  formula = NULL,
  params = UtopParams(),
  variogram = NULL,
  variogram_cloud = NULL,
  variogram_model = NULL,
  predictions = NULL,
  simulations = NULL,
  d_obs = NULL,
  d_pred = NULL,
  g_dist_obs = NULL,
  g_dist_pred = NULL,
  g_dist_pred_obs = NULL,
  g_dist_bin = NULL,
  d_bin = NULL,
  overlap_obs = NULL,
  overlap_pred_obs = NULL,
  var_mat_obs = NULL,
  var_mat_pred_obs = NULL,
  var_mat_pred = NULL,
  var_fit = NULL,
  cv_info = NULL,
  weight = NULL,
  removed = NULL,
  uk_residual = NULL,
  check_vario = NULL
)
}
\arguments{
\item{model}{Variogram model name.}

\item{nugget, unc, cloud, cv}{Logical parameters.}

\item{r_resol, h_resol, amul, dmul, n_max, n_clus}{Numeric or integer tuning
parameters.}

\item{rs_type, hs_type, clus_type, outfile, wlim_method, uk_trend_support}{Character parameters.}

\item{fit_method, g_dist_est, var_clean, max_dist}{Variogram and distance
parameters.}

\item{g_dist_pred}{Geostatistical distance parameter or computed distance
object, depending on the class.}

\item{cn_areas}{Number of areas used in cluster-related calculations.}

\item{partial_overlap}{Whether partially overlapping areas are supported.}

\item{wlim, singular_solve, debug_level}{Additional tuning parameters.}

\item{par_init}{Initial parameter bounds.}

\item{params}{Numeric variogram model parameters.}

\item{ss_err, criterion}{Variogram fitting diagnostics.}

\item{data}{Data stored in variogram objects.}

\item{observations, prediction_locations}{Spatial observations and prediction
locations.}

\item{formula}{Model formula.}

\item{variogram, variogram_cloud, variogram_model}{Variogram results.}

\item{predictions, simulations}{Prediction and simulation results.}

\item{d_obs, d_pred, g_dist_obs, g_dist_pred_obs, g_dist_bin, d_bin}{Discretisation and geostatistical distance objects.}

\item{overlap_obs, overlap_pred_obs}{Overlap matrices.}

\item{var_mat_obs, var_mat_pred_obs, var_mat_pred}{Covariance matrices.}

\item{var_fit, cv_info, weight, removed, uk_residual, check_vario}{Diagnostics
and intermediate results.}
}
\description{
These classes provide the formal object model used by the snake_case utop
API. Prefer \code{utop_object()} and related helpers for user-facing workflows.
}
