% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-check-variogram.R
\name{utop_check_variogram}
\alias{utop_check_variogram}
\title{Create diagnostic variogram checks}
\usage{
utop_check_variogram(object, ...)
}
\arguments{
\item{object}{A \link{Utop} object or \link{UtopVariogramModel}.}

\item{...}{Method-specific arguments. For \link{Utop} objects, supply \code{params}
as a \link{UtopParams} object.}
}
\value{
An updated diagnostic object.
}
\description{
Create diagnostic variogram checks
}
