% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-api.R
\name{utop_fit_variogram}
\alias{utop_fit_variogram}
\title{Fit a variogram model}
\usage{
utop_fit_variogram(object, ...)
}
\arguments{
\item{object}{A \link{Utop} object.}

\item{...}{Arguments passed to the variogram fitting implementation.}
}
\value{
A \link{Utop} object with a fitted variogram model attached.
}
\description{
Fit a variogram model
}
