% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-fit-variogram.R
\name{utop_fit_variogram}
\alias{utop_fit_variogram}
\title{Fit a variogram model}
\usage{
utop_fit_variogram(object, ...)
}
\arguments{
\item{object}{A \link{Utop}, \link{UtopVariogram}, or \link{UtopVariogramCloud} object.}

\item{...}{Method-specific arguments. For \link{Utop} objects, \code{params} and
\code{iprint} control parameter updates and optimiser verbosity.}
}
\value{
A fitted object.
}
\description{
Fit a variogram model
}
