% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-g-dist.R
\name{utop_g_dist}
\alias{utop_g_dist}
\title{Compute geostatistical distances}
\usage{
utop_g_dist(object, ...)
}
\arguments{
\item{object}{A \link{Utop} object, spatial object, or discretisation list.}

\item{...}{Method-specific arguments. For \link{Utop} objects and spatial
inputs, supply \code{params} as a \link{UtopParams} object.}
}
\value{
Geostatistical distances.
}
\description{
Compute geostatistical distances
}
