% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-api.R
\name{utop_params}
\alias{utop_params}
\title{Create utop parameters}
\usage{
utop_params(
  params = list(),
  new_params = list(),
  observations = NULL,
  formula = NULL,
  ...
)
}
\arguments{
\item{params}{Existing parameters as a list or \link{UtopParams} object.}

\item{new_params}{Parameter values to update.}

\item{observations}{Optional observations used to infer starting values.}

\item{formula}{Optional model formula.}

\item{...}{Individual parameter values.}
}
\value{
A \link{UtopParams} object.
}
\description{
Create utop parameters
}
