% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-api.R
\name{utop_variogram}
\alias{utop_variogram}
\title{Create a sample variogram}
\usage{
utop_variogram(object, ...)
}
\arguments{
\item{object}{A \link{Utop} object.}

\item{...}{Arguments passed to the variogram implementation.}
}
\value{
A \link{Utop} object with a variogram attached.
}
\description{
Create a sample variogram
}
