% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-variogram-model.R
\name{utop_variogram_model}
\alias{utop_variogram_model}
\title{Create or update a utop variogram model}
\usage{
utop_variogram_model(
  model = "Ex1",
  sill = NULL,
  range = NULL,
  exp = NULL,
  nugget = NULL,
  exp0 = NULL,
  observations = NULL,
  formula = obs ~ 1
)
}
\arguments{
\item{model}{Variogram model name.}

\item{sill, range, exp, nugget, exp0}{Variogram model parameters.}

\item{observations}{Optional observations used to infer missing parameters.}

\item{formula}{Model formula.}
}
\value{
A \link{UtopVariogramModel} object.
}
\description{
Create or update a utop variogram model
}
